Separate common parts of code into one class ### What's done: Separate common parts of code into one class
