+++ save-core/src/commonMain/kotlin/org/cqfn/save/core/utils/ProcessBuilder.kt
@@ -4,7 +4,10 @@
 
 package org.cqfn.save.core.utils
 
+import okio.FileSystem
 import okio.Path
+import org.cqfn.save.core.logging.logDebug
+import org.cqfn.save.core.logging.logWarn
 
 /**
  * A class that is capable of executing OS processes and returning their output.
@@ -19,6 +22,74 @@ expect class ProcessBuilder() {
      * @return [ExecutionResult] built from process output
      */
     fun exec(command: List<String>, redirectTo: Path?): ExecutionResult
+}
+
+/**
+ * Class contains common fields for all platforms
+ */
+class ProcessBuilderInternal {
+    // Temporary files for stderr and stdout (popen can't separate streams, so we do it ourselves)
+    val fs = FileSystem.SYSTEM
+    val tmpDir = (FileSystem.SYSTEM_TEMPORARY_DIRECTORY / this::class.simpleName!!).also {
+        fs.createDirectory(it)
+    }
+    val stdoutFile = tmpDir / "stdout.txt"
+    val stderrFile = tmpDir / "stderr.txt"
+
+    /**
+     *  Read data from stdout file, we will use it in [ExecutionResult]
+     */
+    fun getStdout(): List<String> {
+        val stdout = fs.read(stdoutFile) {
+            generateSequence { readUtf8Line() }.toList()
+        }
+        return stdout
+    }
+
+    /**
+     * Read data from stderr file, we will use it in [ExecutionResult]
+     */
+    fun getStderr(): List<String> {
+        val stderr = fs.read(stderrFile) {
+            generateSequence { readUtf8Line() }.toList()
+        }
+        return stderr
+    }
+
+    /**
+     * Modify execution command for popen,
+     * stderr will be redirected to tmp file
+     */
+    fun prepare(command: List<String>): String {
+        logDebug("Created file for stderr: ${stderrFile}")
+        val cmd = command.joinToString(" ") + " 2>${stderrFile}"
+        logDebug("Executing: $cmd")
+        return cmd
+    }
+
+    /**
+     * Finish execution and return depends of status and errors
+     */
+    fun logAndReturn(stdout: String, status: Int, redirectTo: Path?): ExecutionResult {
+        if (status == -1) {
+            fs.deleteRecursively(tmpDir)
+            error("Couldn't close the pipe, exit status: $status")
+        }
+        val stderr = getStderr()
+        fs.deleteRecursively(tmpDir)
+        if (stderr.isNotEmpty()) {
+            logWarn(stderr.joinToString("\n"))
+            return ExecutionResult(status, emptyList(), stderr)
+        }
+        if (redirectTo != null) {
+            fs.write(redirectTo) {
+                write(stdout.encodeToByteArray())
+            }
+        } else {
+            logDebug("Execution output:\n${stdout}")
+        }
+        return ExecutionResult(0, stdout.split("\n"), emptyList())
+    }
 }
 
 /**
