+++ save-core/src/jvmMain/kotlin/org/cqfn/save/core/utils/ProcessBuilder.kt
@@ -1,8 +1,6 @@
 package org.cqfn.save.core.utils
 
-import okio.FileSystem
 import okio.Path
-import org.cqfn.save.core.files.createFile
 import org.cqfn.save.core.logging.logDebug
 import org.cqfn.save.core.logging.logWarn
 import java.io.File
@@ -16,13 +14,9 @@ actual class ProcessBuilder {
     private val pb = ProcessBuilder()
 
     actual fun exec(command: List<String>, redirectTo: Path?): ExecutionResult {
-        val fs = FileSystem.SYSTEM
-        val tmpDir = (FileSystem.SYSTEM_TEMPORARY_DIRECTORY / this::class.simpleName!!).also {
-            fs.createDirectory(it)
-        }
-        val stdoutFile = tmpDir / "stdout.txt"
-        val stderrFile = tmpDir / "stderr.txt"
-        logDebug("Created file for stderr: $stderrFile")
+        val common = ProcessBuilderInternal()
+
+        logDebug("Created file for stderr: ${common.stderrFile}")
         val shell = if (System.getProperty("os.name").startsWith("Windows", ignoreCase = true)) listOf("CMD", "/C") else listOf("sh", "-c")
         val cmd = shell + listOf(command.joinToString(" "))
         logDebug("Executing: ${cmd.joinToString(" ")}")
@@ -31,26 +25,22 @@ actual class ProcessBuilder {
                 if (redirectTo != null) {
                     builder.redirectOutput(File(redirectTo.name))
                 } else {
-                    builder.redirectOutput(File(stdoutFile.toString()))
+                    builder.redirectOutput(File(common.stdoutFile.toString()))
                 }
             }
-            .redirectError(File(stderrFile.toString()))
+            .redirectError(File(common.stderrFile.toString()))
             .start()
             .waitFor()
-        val stderr = fs.read(stderrFile) {
-            generateSequence { readUtf8Line() }.toList()
-        }
+        val stdout = common.getStdout()
+        val stderr = common.getStderr()
+        common.fs.deleteRecursively(common.tmpDir)
         if (stderr.isNotEmpty()) {
             logWarn(stderr.joinToString("\n"))
-            fs.deleteRecursively(tmpDir)
             return ExecutionResult(code, emptyList(), stderr)
         }
-        val stdout = fs.read(stdoutFile) {
-            generateSequence { readUtf8Line() }.toList()
+        if (redirectTo == null) {
+            logDebug("Execution output:\n${stdout.joinToString("\n")}")
         }
-        logDebug("Execution output:\n${stdout.joinToString("\n")}")
-
-        fs.deleteRecursively(tmpDir)
         return ExecutionResult(code, redirectTo?.let { File(it.name).readLines() } ?: stdout, emptyList())
     }
 }
