+++ save-core/src/linuxX64Main/kotlin/org/cqfn/save/core/utils/ProcessBuilder.kt
@@ -2,7 +2,6 @@
 
 package org.cqfn.save.core.utils
 
-import okio.FileSystem
 import okio.Path
 import platform.posix.fgets
 import platform.posix.pclose
@@ -10,9 +9,9 @@ import platform.posix.popen
 
 import kotlinx.cinterop.refTo
 import kotlinx.cinterop.toKString
-import org.cqfn.save.core.files.createFile
 import org.cqfn.save.core.logging.logDebug
 import org.cqfn.save.core.logging.logWarn
+import platform.posix.pread
 
 @Suppress("MISSING_KDOC_TOP_LEVEL",
     "MISSING_KDOC_CLASS_ELEMENTS",
@@ -20,15 +19,9 @@ import org.cqfn.save.core.logging.logWarn
 )
 actual class ProcessBuilder {
     actual fun exec(command: List<String>, redirectTo: Path?): ExecutionResult {
-        val fs = FileSystem.SYSTEM
-        val tmpDir = (FileSystem.SYSTEM_TEMPORARY_DIRECTORY / this::class.simpleName!!).also {
-            fs.createDirectory(it)
-        }
-        val stderrFile = tmpDir / "stderr.txt"
-        fs.createFile(stderrFile)
-        logDebug("Created file for stderr: $stderrFile")
-        val cmd = command.joinToString(" ") + " 2>$stderrFile"
-        logDebug("Executing: $cmd")
+        val common = ProcessBuilderInternal()
+        val cmd = common.prepare(command)
+
         val pd = popen(cmd, "r")
             ?: error("Pipe error. Couldn't execute command: `$command`")
         val stdout = buildString {
@@ -37,27 +30,8 @@ actual class ProcessBuilder {
                 append(buffer.toKString())
             }
         }
-
         val status = pclose(pd)
-        if (status == -1) {
-            fs.deleteRecursively(tmpDir)
-            error("Couldn't close the pipe, exit status: $status")
-        }
-        val stderr = fs.read(stderrFile) {
-            generateSequence { readUtf8Line() }.toList()
-        }
-        fs.deleteRecursively(tmpDir)
-        if (stderr.isNotEmpty()) {
-            logWarn(stderr.joinToString("\n"))
-            return ExecutionResult(status, emptyList(), stderr)
-        }
-        if (redirectTo != null) {
-            fs.write(redirectTo) {
-                write(stdout.encodeToByteArray())
-            }
-        } else {
-            logDebug("Execution output:\n${stdout}")
-        }
-        return ExecutionResult(0, stdout.split("\n"), emptyList())
+
+        return common.logAndReturn(stdout, status, redirectTo)
     }
 }
